THL_TB_Data_Analysis.md
In this Rmd File, we provide the detailed  examination  on the tuberculosis (TB) data for unravelling insights from the collected THL data in their repository. 

We discuss on time series, autoregressive (AR) and vector autoregressive (VAR) models to model the national data on TB for Finland between two subtypes disease namely pulmonary TB and other organs TB. We found the temporal process of these subtypes are trend stationary and corresponding to each other in Finland due to the co-movement between them.

Finally, the joint forecast between the subtypes disease depict that the number of cases for both will decline for the period 2025-2029 in Finland.